requirements.txt
#python>=3.9,<3.12
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
lightgbm==4.0.0
imbalanced-learn==0.11.0
mlflow==2.7.1
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
python-multipart==0.0.6
python-dotenv==1.0.0
pyyaml==6.0.1
joblib==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1
streamlit==1.28.0
streamlit-aggrid==0.3.4
evidently==0.4.33
prometheus-client==0.17.1
prefect>=2.0,<3.0
locust==2.15.1
pytest==7.4.2
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.7.0
flake8==6.1.0
mypy==1.5.1
pre-commit==3.4.0
plotly-express==0.4.1
